feat: 회원가입 시 이미지 추가

+++ src/main/kotlin/udumeoli/tripphoto/user/dto/UpdateProfileInput.kt
@@ -2,5 +2,5 @@ package udumeoli.tripphoto.user.dto
 
 data class UpdateProfileInput(
     val nickname: String,
-    val profileImage: Int?,
+    val profileImage: Long?,
 )
